# openapi3_1.yaml
openapi: '3.1.0'
info:
  version: '1.0.0'
  title: 'Food Driver API'
  description: System obsługi zamówień online dla restauracji
servers:
  - url: http://localhost:3000/api
    description: Środowisko deweloperskie
paths:
  /users/sign-up:
    post:
      operationId: SignUp
      description: Utworzenie konta użytkownika
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/Name'
                surname:
                  $ref: '#/components/schemas/Name'
                email:
                  $ref: '#/components/schemas/Email'
                confirmEmail:
                  $ref: '#/components/schemas/Email'
                password:
                  $ref: '#/components/schemas/Password'
                confirmPassword:
                  $ref: '#/components/schemas/Password'
              required:
                - name
                - surname
                - email
                - confirmEmail
                - password
                - confirmPassword
              additionalProperties: false
      responses:
        '201':
          $ref: '#/components/responses/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/activate:
    post:
      operationId: ActivateAccount
      description: Aktywacja konta użytkownika
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  $ref: '#/components/schemas/JWT'
              required:
                - token
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/reset-password:
    post:
      operationId: ResetPassword
      description: Reset hasła użytkownika
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
              required:
                - email
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/reset-password/confirm:
    post:
      operationId: ResetPasswordConfirm
      description: Potwierdzenie resetu hasła użytkownika
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  $ref: '#/components/schemas/JWT'
                password:
                  $ref: '#/components/schemas/Password'
                confirmPassword:
                  $ref: '#/components/schemas/Password'
              required:
                - token
                - password
                - confirmPassword
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/sign-in:
    post:
      operationId: SignIn
      description: Logowanie na konto użytkownika
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  $ref: '#/components/schemas/Password'
              required:
                - email
                - password
              additionalProperties: false
      responses:
        '200':
          description: Para tokenów autentykacji i odświeżania
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/refresh:
    post:
      operationId: RefreshTokens
      description: Odświeżanie tokenów autentykacji i odświeżania
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  $ref: '#/components/schemas/JWT'
              required:
                - refreshToken
              additionalProperties: false
      responses:
        '200':
          description: Para tokenów autentykacji i odświeżania
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/me:
    get:
      operationId: GetMe
      description: Pobieranie danych zalogowanego użytkownika
      security:
        - bearerAuth: [Klient]
      responses:
        '200':
          description: Dane użytkownika
          content:
            application/json:
              schema:
                type: object
                properties:
                  idUser:
                    $ref: '#/components/schemas/Id'
                  name:
                    $ref: '#/components/schemas/Name'
                  surname:
                    $ref: '#/components/schemas/Name'
                  email:
                    $ref: '#/components/schemas/Email'
                  street:
                    anyOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/Name'
                  building:
                    anyOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/Name'
                  apartment:
                    anyOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/Name'
                  postalCode:
                    anyOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/PostalCode'
                  city:
                    anyOf:
                      - type: 'null'
                      - $ref: '#/components/schemas/Name'
                  role:
                    $ref: '#/components/schemas/UserRole'
                  created:
                    $ref: '#/components/schemas/DateTime'
                required:
                  - idUser
                  - name
                  - surname
                  - email
                  - street
                  - building
                  - apartment
                  - postalCode
                  - city
                  - role
                  - created
                additionalProperties: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
    put:
      operationId: UpdateMe
      description: Edycja danych zalogowanego użytkownika
      security:
        - bearerAuth: [Klient]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/Name'
                surname:
                  $ref: '#/components/schemas/Name'
                street:
                  $ref: '#/components/schemas/Name'
                building:
                  $ref: '#/components/schemas/Name'
                apartment:
                  $ref: '#/components/schemas/Name'
                postalCode:
                  $ref: '#/components/schemas/PostalCode'
                city:
                  $ref: '#/components/schemas/Name'
              required:
                - name
                - surname
                - street
                - building
                - apartment
                - postalCode
                - city
              additionalProperties: false
      responses:
        '204':
          description: Pomyślnie edytowano dane użytkownika
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/me/subscribe:
    post:
      operationId: SubscribePushNotifications
      description: Dodanie subskrypcji powiadomień Push
      security:
        - bearerAuth: [Klient]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '204':
          description: Pomyślnie dodano subskrypcję powiadomień Push
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
  /users/me/change-password:
    post:
      operationId: ChangePassword
      description: Zmiana hasła do konta użytkownika
      security:
        - bearerAuth: [Klient]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  $ref: '#/components/schemas/Password'
                newPassword:
                  $ref: '#/components/schemas/Password'
                confirmNewPassword:
                  $ref: '#/components/schemas/Password'
              required:
                - currentPassword
                - newPassword
                - confirmNewPassword
              additionalProperties: false
      responses:
        '204':
          description: Pomyślnie zmieniono hasło do konta użytkownika
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
...
  /users/me/restaurant/menu-items:
    post:
      operationId: CreateMeRestaurantMenuItem
      description: Utworzenie pozycji menu
      security:
        - bearerAuth: [Restaurator]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/Name'
                description:
                  $ref: '#/components/schemas/Description'
                price:
                  $ref: '#/components/schemas/Price'
                idImage:
                  $ref: '#/components/schemas/Id'
              required:
                - name
                - description
                - price
                - idImage
              additionalProperties: false
      responses:
        '204':
          description: Pomyślnie utworzono pozycję menu
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Users
...
components:
  responses:
    Message:
      description: Odpowiedź serwera z komunikatem dla użytkownika
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Info'
    BadRequest:
      description: Nieprawidłowa treść żądania
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Nieprawidłowy token autentykacji
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Wystąpił wewnętrzny błąd serwera
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Id:
      description: Unikalny identyfikator
      type: integer
    JWT:
      description: JSON Web token
      type: string
      minLength: 1
    Name:
      type: string
      minLength: 1
      maxLength: 200
    Description:
      type: string
      minLength: 1
      maxLength: 3000
...
    Email:
      type: string
      format: email
      maxLength: 320
    Password:
      type: string
      minLength: 8
    Price:
      type: string
      format: ^\d+(.\d{1,2})?$
    PostalCode:
      type: string
      pattern: ^\d{2}-\d{3}$
    DateTime:
      type: string
      format: date-time
    UserRole:
      type: string
      enum:
        - Klient
        - Restaurator
        - Dostawca
        - Administrator
...
    TokensPair:
      type: object
      properties:
        authenticationToken:
          $ref: '#/components/schemas/JWT'
        refreshToken:
          $ref: '#/components/schemas/JWT'
      required:
        - authenticationToken
        - refreshToken
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      description: Token na okaziciela JWT
      type: http
      scheme: bearer
      bearerFormat: JWT